+++ plugins/modules/ah_approval.py
@@ -41,13 +41,13 @@ options:
       required: True
       type: str
 
-extends_documentation_fragment: ansible.automation_hub.auth
+extends_documentation_fragment: galaxy.galaxy.auth
 """
 
 
 EXAMPLES = """
 - name: Approve namespace.collection_name:v1.0.0
-  ansible.automation_hub.ah_approval:
+  galaxy.galaxy.ah_approval:
     namespace: namespace
     name: collection_name
     version: v1.0.0