+++ plugins/modules/ah_ee_namespace.py
@@ -33,7 +33,7 @@ deprecated:
   why: The endpoint has been removed and is not supported in AAP 2.3 onwards
   alternative: None
   removed_in: 3.0.0
-  removed_from_collection: ansible.automation_hub
+  removed_from_collection: galaxy.galaxy
 options:
   name:
     description:
@@ -71,12 +71,12 @@ notes:
   - When the module creates a namespace, the private automation hub web UI does not display that new namespace.
     You must first create a repository in the namespace, with C(podman push) for example, for the web UI to show the namespace.
   - When the module grants access to a group, the permissions associated to that group apply.
-extends_documentation_fragment: ansible.automation_hub.auth_ui
+extends_documentation_fragment: galaxy.galaxy.auth_ui
 """
 
 EXAMPLES = r"""
 - name: Ensure the namespace exists
-  ansible.automation_hub.ah_ee_namespace:
+  galaxy.galaxy.ah_ee_namespace:
     name: ansible-automation-platform-20-early-access
     state: present
     ah_host: hub.example.com
@@ -84,7 +84,7 @@ EXAMPLES = r"""
     ah_password: Sup3r53cr3t
 
 - name: Ensure the namespace has a new name
-  ansible.automation_hub.ah_ee_namespace:
+  galaxy.galaxy.ah_ee_namespace:
     name: ansible-automation-platform-20-early-access
     new_name: custom-ee-01
     state: present
@@ -93,7 +93,7 @@ EXAMPLES = r"""
     ah_password: Sup3r53cr3t
 
 - name: Ensure the namespace is removed
-  ansible.automation_hub.ah_ee_namespace:
+  galaxy.galaxy.ah_ee_namespace:
     name: custom-ee-01
     state: absent
     ah_host: hub.example.com
@@ -101,7 +101,7 @@ EXAMPLES = r"""
     ah_password: Sup3r53cr3t
 
 - name: Ensure only the operators group can manage the namespace
-  ansible.automation_hub.ah_ee_namespace:
+  galaxy.galaxy.ah_ee_namespace:
     name: ansible-automation-platform-20-early-access
     state: present
     groups:
@@ -112,7 +112,7 @@ EXAMPLES = r"""
     ah_password: Sup3r53cr3t
 
 - name: Ensure the managers group can also manage the namespace
-  ansible.automation_hub.ah_ee_namespace:
+  galaxy.galaxy.ah_ee_namespace:
     name: ansible-automation-platform-20-early-access
     state: present
     groups: