+++ plugins/modules/ah_ee_registry.py
@@ -107,12 +107,12 @@ options:
 notes:
   - Supports C(check_mode).
   - Only works with private automation hub v4.4.0 or later.
-extends_documentation_fragment: ansible.automation_hub.auth_ui
+extends_documentation_fragment: galaxy.galaxy.auth_ui
 """
 
 EXAMPLES = r"""
 - name: Add a remote registry to AH
-  ansible.automation_hub.ah_ee_registry:
+  galaxy.galaxy.ah_ee_registry:
     name: my_quayio
     state: present
     url: https://quay.io/my/registry
@@ -121,7 +121,7 @@ EXAMPLES = r"""
     ah_password: Sup3r53cr3t
 
 - name: Add a remote registry which requires auth to AH
-  ansible.automation_hub.ah_ee_registry:
+  galaxy.galaxy.ah_ee_registry:
     name: my_quayio_auth
     state: present
     url: https://quay.io/my/registry
@@ -132,7 +132,7 @@ EXAMPLES = r"""
     ah_password: Sup3r53cr3t
 
 - name: Remove a remote registry from AH
-  ansible.automation_hub.ah_ee_registry:
+  galaxy.galaxy.ah_ee_registry:
     name: examplehub
     state: absent
     ah_host: hub.example.com