+++ plugins/modules/ah_group_perm.py
@@ -30,7 +30,7 @@ deprecated:
   why: The endpoint has been removed and is not supported in AAP 2.3 onwards. It is replaaced by role permisions.
   alternative: None
   removed_in: 3.0.0
-  removed_from_collection: ansible.automation_hub
+  removed_from_collection: galaxy.galaxy
 options:
   name:
     description:
@@ -65,8 +65,8 @@ options:
     default: present
     choices: [absent, present]
 seealso:
-  - module: ansible.automation_hub.ah_group
-  - module: ansible.automation_hub.ah_user
+  - module: galaxy.galaxy.ah_group
+  - module: galaxy.galaxy.ah_user
 notes:
   - Supports C(check_mode).
   - This module only works up to Automation Hub version 4.5 (AAP 2.2)
@@ -74,13 +74,13 @@ notes:
     C(change_container), C(change_image_tag), C(create_container), and
     C(push_container)) are only available with private automation hub v4.3.2 or
     later.
-extends_documentation_fragment: ansible.automation_hub.auth_ui
+extends_documentation_fragment: galaxy.galaxy.auth_ui
 """
 
 
 EXAMPLES = r"""
 - name: Ensure the operators have the correct permissions to manage users
-  ansible.automation_hub.ah_group_perm:
+  galaxy.galaxy.ah_group_perm:
     name: operators
     perms:
       - add_user
@@ -93,7 +93,7 @@ EXAMPLES = r"""
     ah_password: Sup3r53cr3t
 
 - name: Ensure the administrators have all the permissions
-  ansible.automation_hub.ah_group_perm:
+  galaxy.galaxy.ah_group_perm:
     name: administrators
     perms: "*"
     state: present
@@ -102,7 +102,7 @@ EXAMPLES = r"""
     ah_password: Sup3r53cr3t
 
 - name: Ensure the developers cannot manage groups nor users
-  ansible.automation_hub.ah_group_perm:
+  galaxy.galaxy.ah_group_perm:
     name: developers
     perms:
       - add_user