+++ plugins/modules/ah_role.py
@@ -63,18 +63,18 @@ options:
     default: present
     choices: [absent, present]
 seealso:
-  - module: ansible.automation_hub.ah_group
-  - module: ansible.automation_hub.ah_user
+  - module: galaxy.galaxy.ah_group
+  - module: galaxy.galaxy.ah_user
 notes:
   - Supports C(check_mode).
   - This module only works up to Automation Hub version 4.6 (AAP 2.3)
-extends_documentation_fragment: ansible.automation_hub.auth_ui
+extends_documentation_fragment: galaxy.galaxy.auth_ui
 """
 
 
 EXAMPLES = r"""
 - name: Ensure the operators have the correct permissions to manage users
-  ansible.automation_hub.ah_role:
+  galaxy.galaxy.ah_role:
     name: galaxy.operators
     perms:
       - add_user
@@ -87,7 +87,7 @@ EXAMPLES = r"""
     ah_password: Sup3r53cr3t
 
 - name: Ensure the administrators have all the permissions
-  ansible.automation_hub.ah_role:
+  galaxy.galaxy.ah_role:
     name: galaxy.administrators
     perms: "*"
     state: present
@@ -96,7 +96,7 @@ EXAMPLES = r"""
     ah_password: Sup3r53cr3t
 
 - name: Ensure the developers cannot manage groups nor users
-  ansible.automation_hub.ah_role:
+  galaxy.galaxy.ah_role:
     name: galaxy.developers
     perms:
       - add_user